Refactored RouteMap
