+++ frontend/boutique_app/src/components/RouteMap.vue
@@ -1,4 +1,8 @@
 <template>
+  <div v-if="totalTravelTime" class="travel-time">
+    <h2>Total Travel Time: {{ totalTravelTime }}</h2>
+  </div>
+
   <GMapMap
       :center="center"
       :zoom="zoom"
@@ -15,6 +19,19 @@
         :path="routePath"
         :options="{ strokeColor: '#FF0000', strokeOpacity: 1.0, strokeWeight: 2 }"
     />
+
+    <div v-if="transitDetails.length" class="transit-details">
+      <h2>Transit Details</h2>
+      <ul>
+        <li v-for="(detail, index) in transitDetails" :key="index">
+          <p><strong>Line Number:</strong> {{ detail.lineNumber }}</p>
+          <p><strong>Line:</strong> {{ detail.lineName }}</p>
+          <p><strong>Vehicle:</strong> {{ detail.vehicleType }}</p>
+          <p><strong>Departure Stop:</strong> {{ detail.departureStop }}</p>
+          <p><strong>Arrival Stop:</strong> {{ detail.arrivalStop }}</p>
+        </li>
+      </ul>
+    </div>
   </GMapMap>
 </template>
 
@@ -27,53 +44,66 @@ export default defineComponent({
     startAddress: {
       type: String,
       required: true,
+    },
+    travelMode: {
+      type: String,
+      required: true,
     }
   },
   setup(props) {
     const center = ref({ lat: 48.210033, lng: 16.363449 });
-    console.log("Start address:", props.startAddress);
-
     const zoom = ref(12);
-    const markers = ref<any[]>([]);
-    const routePath = ref<any[]>([]);
+    const markers = ref([]);
+    const routePath = ref([]);
+    const transitDetails = ref([]);
+    const totalTravelTime = ref('');
 
     const onMapLoaded = () => {
-      const geocoder = new google.maps.Geocoder();
       const directionsService = new google.maps.DirectionsService();
-      const endAddress = 'Höchstädtpl. 6, 1200 Wien';
+      let travelMode = google.maps.TravelMode.DRIVING;
+      if (props.travelMode === 'train') {
+        travelMode = google.maps.TravelMode.TRANSIT;
+      }
 
-      geocoder.geocode({ address: props.startAddress }, (startResults, startStatus) => {
-        if (startStatus === google.maps.GeocoderStatus.OK && startResults[0]) {
-          const startLocation = startResults[0].geometry.location;
-          markers.value.push({ position: startLocation });
+      const request = {
+        origin: props.startAddress,
+        destination: 'Höchstädtpl. 6, 1200 Wien',
+        travelMode: travelMode,
+      };
 
-          geocoder.geocode({ address: endAddress }, (endResults, endStatus) => {
-            if (endStatus === google.maps.GeocoderStatus.OK && endResults[0]) {
-              const endLocation = endResults[0].geometry.location;
-              markers.value.push({ position: endLocation });
+      directionsService.route(request, (result, status) => {
+        if (status === google.maps.DirectionsStatus.OK && result.routes[0]) {
+          const overviewPath = result.routes[0].overview_path;
+          routePath.value = overviewPath.map(point => ({
+            lat: point.lat(),
+            lng: point.lng(),
+          }));
 
-              const request = {
-                origin: startLocation,
-                destination: endLocation,
-                travelMode: google.maps.TravelMode.DRIVING,
-              };
-
-              directionsService.route(request, (result, status) => {
-                if (status === google.maps.DirectionsStatus.OK && result.routes[0]) {
-                  routePath.value = result.routes[0].overview_path.map((point: any) => ({
-                    lat: point.lat(),
-                    lng: point.lng(),
-                  }));
-                } else {
-                  console.error('Error fetching directions', status);
-                }
-              });
-            } else {
-              console.error('End address geocoding failed:', endStatus);
-            }
+          // Add markers for start and end locations
+          markers.value.push({
+            position: result.routes[0].legs[0].start_location,
           });
+          markers.value.push({
+            position: result.routes[0].legs[0].end_location,
+          });
+
+          totalTravelTime.value = result.routes[0].legs[0].duration.text;
+
+          if (props.travelMode === 'train') {
+            // Extract transit details
+            const legs = result.routes[0].legs[0];
+            transitDetails.value = legs.steps
+                .filter(step => step.travel_mode === 'TRANSIT')
+                .map(step => ({
+                  lineName: step.transit.line.name,
+                  lineNumber: step.transit.line.short_name,
+                  vehicleType: step.transit.line.vehicle.type,
+                  departureStop: step.transit.departure_stop.name,
+                  arrivalStop: step.transit.arrival_stop.name,
+                }));
+          }
         } else {
-          console.error('Start address geocoding failed:', startStatus);
+          console.error('Error fetching directions:', status);
         }
       });
     };
@@ -83,6 +113,8 @@ export default defineComponent({
       zoom,
       markers,
       routePath,
+      transitDetails,
+      totalTravelTime,
       onMapLoaded,
     };
   },
@@ -90,4 +122,24 @@ export default defineComponent({
 </script>
 
 <style scoped>
+.transit-details {
+  margin-top: 1rem;
+}
+
+.transit-details h2 {
+  font-size: 1.2rem;
+}
+
+.transit-details ul {
+  list-style-type: none;
+  padding: 0;
+}
+
+.transit-details li {
+  margin-bottom: 1rem;
+}
+
+.transit-details p {
+  margin: 0;
+}
 </style>
